Move AEC, BAC, and frequency scale factor data to RMG-database
