+++ arkane/statmech.py
@@ -1114,86 +1114,6 @@ def project_rotors(conformer, hessian, rotors, linear, is_ts, get_projected_out_
     return np.sqrt(eig[-n_vib:]) / (2 * math.pi * constants.c * 100)
 
 
-def assign_frequency_scale_factor(freq_level):
-    """
-    Assign the frequency scaling factor according to the model chemistry.
-    Refer to https://comp.chem.umn.edu/freqscale/index.html for future updates of these factors
-
-    Sources:
-        [1] I.M. Alecu, J. Zheng, Y. Zhao, D.G. Truhlar, J. Chem. Theory Comput. 2010, 6, 2872, DOI: 10.1021/ct100326h
-        [2] http://cccbdb.nist.gov/vibscalejust.asp
-        [3] http://comp.chem.umn.edu/freqscale/190107_Database_of_Freq_Scale_Factors_v4.pdf
-        [4] Calculated as described in 10.1021/ct100326h
-        [5] J.A. Montgomery, M.J. Frisch, J. Chem. Phys. 1999, 110, 2822–2827, DOI: 10.1063/1.477924
-
-    Args:
-        freq_level (str, unicode): The frequency level of theory.
-
-    Returns:
-        float: The frequency scaling factor (1 by default).
-    """
-    freq_dict = {'hf/sto-3g': 0.817,  # [2]
-                 'hf/6-31g': 0.903,  # [2]
-                 'hf/6-31g(d)': 0.899,  # [2]
-                 'hf/6-31g(d,p)': 0.903,  # [2]
-                 'hf/6-31g+(d,p)': 0.904,  # [2]
-                 'hf/6-31+g(d,p)': 0.915 * 1.014,  # [1] Table 7
-                 'pm3': 0.940 * 1.014,  # [1] Table 7, the 0.940 value is the ZPE scale factor
-                 'pm6': 1.078 * 1.014,  # [1] Table 7, the 1.078 value is the ZPE scale factor
-                 'b3lyp/6-31g(d,p)': 0.961,  # [2]
-                 'b3lyp/6-311g(d,p)': 0.967,  # [2]
-                 'b3lyp/6-311+g(3df,2p)': 0.967,  # [2]
-                 'b3lyp/6-311+g(3df,2pd)': 0.970,  # [2]
-                 'm06-2x/6-31g(d,p)': 0.952,  # [2]
-                 'm06-2x/6-31+g(d,p)': 0.979,  # [3]
-                 'm06-2x/6-311+g(d,p)': 0.983,  # [3]
-                 'm06-2x/6-311++g(d,p)': 0.983,  # [3]
-                 'm06-2x/cc-pvtz': 0.955,  # [2]
-                 'm06-2x/aug-cc-pvdz': 0.993,  # [3]
-                 'm06-2x/aug-cc-pvtz': 0.985,  # [1] Table 3, [3]
-                 'm06-2x/def2-tzvp': 0.984,  # [3]
-                 'm06-2x/def2-qzvp': 0.983,  # [3]
-                 'm06-2x/def2-tzvpp': 0.983,  # [1] Table 3, [3]
-                 'm08so/mg3s*': 0.995,  # [1] Table 3, taken as 'M08-SO/MG3S'
-                 'wb97x-d/aug-cc-pvtz': 0.988,  # [3], taken as 'ωB97X-D/maug-cc-pVTZ'
-                 'wb97xd/6-311++g(d,p)': 0.988,  # [4]
-                 'wb97xd/def2tzvp': 0.988,  # [4]
-                 'wb97xd/def2svp': 0.986,  # [4]
-                 'apfd/def2tzvp': 0.993,  # [4]
-                 'apfd/def2tzvpp': 0.992,  # [4]
-                 'mp2_rmp2_pvdz': 0.953,  # [2], taken as 'MP2/cc-pVDZ'
-                 'mp2_rmp2_pvtz': 0.950,  # [2], taken as 'MP2/cc-pVTZ'
-                 'mp2_rmp2_pvqz': 0.962,  # [2], taken as 'MP2/cc-pVQZ'
-                 'cbs-qb3': 0.99 * 1.014,  # [5], the 0.99 value is the ZPE scale factor of CBS-QB3
-                 'cbs-qb3-paraskevas': 0.99 * 1.014,  # [5], the 0.99 value is the ZPE scale factor of CBS-QB3
-                 'ccsd-f12/cc-pvdz-f12': 0.947,  # [2], taken as 'CCSD/cc-pVDZ'
-                 'ccsd(t)/cc-pvdz': 0.979,  # [2]
-                 'ccsd(t)/cc-pvtz': 0.975,  # [2]
-                 'ccsd(t)/cc-pvqz': 0.970,  # [2]
-                 'ccsd(t)/aug-cc-pvdz': 0.963,  # [2]
-                 'ccsd(t)/aug-cc-pvtz': 1.001,  # [3]
-                 'ccsd(t)/aug-cc-pvqz': 0.975,  # [2]
-                 'ccsd(t)/cc-pv(t+d)z': 0.965,  # [2]
-                 'ccsd(t)-f12/cc-pvdz-f12': 0.997,  # [3], taken as 'CCSD(T)-F12a/cc-pVDZ-F12'
-                 'ccsd(t)-f12/cc-pvtz-f12': 0.998,  # [3], taken as 'CCSD(T)-F12a/cc-pVTZ-F12'
-                 'ccsd(t)-f12/cc-pvqz-f12': 0.998,  # [3], taken as 'CCSD(T)-F12b/VQZF12//CCSD(T)-F12a/TZF'
-                 'ccsd(t)-f12/cc-pcvdz-f12': 0.997,  # [3], taken as 'CCSD(T)-F12a/cc-pVDZ-F12'
-                 'ccsd(t)-f12/cc-pcvtz-f12': 0.998,  # [3], taken as 'CCSD(T)-F12a/cc-pVTZ-F12'
-                 'ccsd(t)-f12/aug-cc-pvdz': 0.997,  # [3], taken as 'CCSD(T)/cc-pVDZ'
-                 'ccsd(t)-f12/aug-cc-pvtz': 0.998,  # [3], taken as CCSD(T)-F12a/cc-pVTZ-F12
-                 'ccsd(t)-f12/aug-cc-pvqz': 0.998,  # [3], taken as 'CCSD(T)-F12b/VQZF12//CCSD(T)-F12a/TZF'
-                 }
-    scaling_factor = freq_dict.get(freq_level.lower(), 1)
-    if scaling_factor == 1:
-        logging.warning('No frequency scaling factor found for model chemistry {0}. Assuming a value of unity. '
-                        'This will affect the partition function and all quantities derived from it '
-                        '(thermo quantities and rate coefficients).'.format(freq_level))
-    else:
-        logging.info('Assigned a frequency scale factor of {0} for the frequency level of theory {1}'.format(
-            scaling_factor, freq_level))
-    return scaling_factor
-
-
 def determine_rotor_symmetry(energies, label, pivots):
     """
     Determine the rotor symmetry number from the potential scan given in :list:`energies` in J/mol units
